:sparkles: feat : GET /user/id 구현 #79
